Accommodate NNS bug (working towards #10)
NNS loads textures and makes accommodations in VRAM allocation for palette data containing palette4 palettes. It fails to check the correct flag, however, checking the flag being set in the tex4x4 resource flags, rather than the palette resource flags. This sets that flag in the tex4x4 section where it will actually be checked erroneously. This doesn't fix #10, but this I believe covers part of the basis of that problem.